chore: Move the monaco editor setup code to its own file
