+++ src/aux-server/aux-web/shared/MonacoHelpers.ts
@@ -1,0 +1,62 @@
+import * as monaco from 'monaco-editor';
+import EditorWorker from 'worker-loader!monaco-editor/esm/vs/editor/editor.worker.js';
+import TypescriptWorker from 'worker-loader!monaco-editor/esm/vs/language/typescript/ts.worker';
+import { calculateFormulaDefinitions } from './FormulaHelpers';
+import { lib_es2015_dts } from 'monaco-editor/esm/vs/language/typescript/lib/lib.js';
+import { SimpleEditorModelResolverService } from 'monaco-editor/esm/vs/editor/standalone/browser/simpleServices';
+
+export function setup() {
+    // Tell monaco how to create the web workers
+    (<any>self).MonacoEnvironment = {
+        getWorker: function(moduleId: string, label: string) {
+            if (label === 'typescript' || label === 'javascript') {
+                return new TypescriptWorker();
+            }
+            return new EditorWorker();
+        },
+    };
+
+    // Set diagnostics
+    monaco.languages.typescript.javascriptDefaults.setDiagnosticsOptions({
+        noSemanticValidation: false,
+        noSyntaxValidation: false,
+    });
+
+    // Set compiler options
+    monaco.languages.typescript.javascriptDefaults.setCompilerOptions({
+        target: monaco.languages.typescript.ScriptTarget.ES2015,
+
+        // Auto-import the given libraries
+        lib: ['defaultLib:lib.es2015.d.ts', 'file:///formula-lib.d.ts'],
+
+        allowJs: true,
+    });
+
+    // Eagerly sync models to get intellisense for all models
+    monaco.languages.typescript.javascriptDefaults.setEagerModelSync(true);
+
+    // Register the ES2015 core library
+    monaco.languages.typescript.javascriptDefaults.addExtraLib(
+        lib_es2015_dts,
+        'defaultLib:lib.es2015.d.ts'
+    );
+
+    // Register the formula library
+    monaco.languages.typescript.javascriptDefaults.addExtraLib(
+        calculateFormulaDefinitions(),
+        'file:///formula-lib.d.ts'
+    );
+
+    /**
+     * Monkeypatch to make 'Find All References' work across multiple files
+     * https://github.com/Microsoft/monaco-editor/issues/779#issuecomment-374258435
+     */
+    SimpleEditorModelResolverService.prototype.findModel = function(
+        editor: monaco.editor.IStandaloneCodeEditor,
+        resource: any
+    ) {
+        return monaco.editor
+            .getModels()
+            .find(model => model.uri.toString() === resource.toString());
+    };
+}
